MscmpSystSession.Types (mscmp_syst_session v0.1.0)

Data types defined by and for use with the McmpSystSession component.

Types

session_data()

@type session_data() :: map()

The session contents, the final data to be stored after it has been built with Plug.Conn.put_session/3 and the other session manipulating functions.

session_name()

@type session_name() :: binary()

The internal reference to the session in the store.

session_params()

@type session_params() :: %{
  optional(:internal_name) => session_name(),
  optional(:session_data) => session_data(),
  optional(:session_expires) => DateTime.t()
}